refactor, remove old push service
