+++ index.js
@@ -6,7 +6,7 @@ import {Alert, AppRegistry, Platform} from 'react-native';
 import App from './App';
 import {name as appName} from './app.json';
 import BackgroundFetch from 'react-native-background-fetch';
-import PushService from './PushService';
+import PushService from './services/PushService';
 import {Config} from './config';
 import {notifications} from './mastoApi/mastoApi';
 import {getSession, updateNotifId} from './services/storage';
@@ -15,6 +15,7 @@ import {getNotifs} from './services/bridgesNotifs';
 this.id = 0;
 AppRegistry.registerComponent(appName, () => App);
 
+// if Android and app is closed getNotifs
 let MyHeadlessTask = async () => {
   getNotifs('INDEX.js').then(
     BackgroundFetch.finish('com.transistorsoft.fetch'),
